Bird is flying at least

+++ KanskeKulSpel/src/Game/Entities/MovingEntity.h
@@ -18,7 +18,7 @@ protected:
     virtual void handleCollision(const Entity* collider);
     void addCollisionMomentum(sf::Vector2f colliderMomentum, sf::Vector2f colliderPos, float colliderMass);
 
-    sf::Vector2i acceleration;
+    sf::Vector2f acceleration;
     sf::Vector2f momentum;
 
     bool addedMomentum;
